+++ bot/Utils/PrintLogs/__init__.py
@@ -73,7 +73,7 @@ class PrintBot(CrawJUD):
     def file_log(cls, self: Self) -> None:
         """Write log messages to a file."""
         try:
-            savelog = os.path.join(pathlib.Path(__file__).cwd(), "exec", self.pid, f"LogFile - PID {self.pid}.txt")
+            savelog = os.path.join(pathlib.Path(__file__).cwd(), "temp", self.pid, f"LogFile - PID {self.pid}.txt")
             with open(savelog, "a") as f:
                 for mensagem in self.list_messages:
                     if self.pid in mensagem:
