+++ bot/scripts/elaw/cadastro.py
@@ -1049,7 +1049,7 @@ class Cadastro(CrawJUD):
 
         """
         name_comprovante = f"Comprovante Cadastro - {self.bot_data.get('NUMERO_PROCESSO')} - PID {self.pid}.png"
-        savecomprovante = os.path.join(pathlib.Path(__file__).cwd(), "exec", self.pid, name_comprovante)
+        savecomprovante = os.path.join(pathlib.Path(__file__).cwd(), "temp", self.pid, name_comprovante)
         self.driver.get_screenshot_as_file(savecomprovante)
         self.append_success([self.bot_data.get("NUMERO_PROCESSO"), name_comprovante, self.pid])
