+++ utils/make_celery.py
@@ -1,0 +1,25 @@
+"""Celery configuration for Quart application."""
+
+from celery import Celery
+from quart import Quart
+
+
+def make_celery(app: Quart) -> Celery:
+    """Create and configure a Celery instance with Quart application context.
+
+    Args:
+        app (Quart): The Quart application instance.
+
+    Returns:
+        Celery: Configured Celery instance.
+
+    """
+    celery = Celery(app.import_name)
+    celery.conf.update(app.config["CELERY"])
+
+    class ContextTask(celery.Task):
+        def __call__(self, *args: tuple, **kwargs: dict) -> any:  # -> any:
+            return self.run(*args, **kwargs)
+
+    celery.Task = ContextTask
+    return celery
